refactor: moved client code to separate package
